frontend: Further separate out components into their own files
